Project 0x0F - Retrying Tasks 1, 2 and 6 --> 13

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 """Deletes all State objects with a name containing the letter 'a',
 from the database hbtn_0e_6_usa"""
-import sys
+from sys import argv
 
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
@@ -9,15 +9,14 @@ from sqlalchemy.orm import sessionmaker
 from model_state import Base, State
 
 if __name__ == "__main__":
-    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
-                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
-                           pool_pre_ping=True)
+    eng = f'mysql+mysqldb://{argv[1]}:{argv[2]}@localhost/{argv[3]}'
+    engine = create_engine(eng, pool_pre_ping=True)
+    Base.metadata.create_all(engine)
     Session = sessionmaker(bind=engine)
+
     session = Session()
-
-    states = session.query(State).filter(State.name.like(r'%a%')).all()
-
-    for state in states:
-        session.delete(state)
-
+    to_delete = session.query(State).filter(State.name.like('%a%')).all()
+    for state in to_delete:
+        if 'a' in state.name:
+            session.delete(state)
     session.commit()
